Rename EntityCollectionRelationship to EntityNavigationCollection
